Added an integration for `ptpython` referenced as `experimental editor`.
This integration is not perfect. I had to read the configuration file
and ended up duplication some classes and functions from `sardine` to
`fishery` in order to do it.

There is a nasty bug that shows up when quitting from the experimental editor.
This issue is already documented by `ptpython` and should have been fixed
already if I am to believe GitHub issues:
https://github.com/prompt-toolkit/python-prompt-toolkit/pull/1457.

This is probably worth reporting to the maintainers!