Ableton Link initial integration test and other changes: - Sardine clock state is now forwarded through OSC to port 12345
I wasn't initially down for implementing Ableton Link support in
Sardine at such an early stage (one more thing to support). However, it
looks like it's not so difficult to achieve using the LinkPython
package. This is really experimental and I don't expect it to work that
easily but it's enough for some local network synchronisation for now.