Mockup of the new lexer/parser system
This is a mockup of the new and improved parser. It is now able to
handle more operations than before. It handles whitespaces properly. I included
new tokens (time, phase, unix time, measure) and some new operations (such as
the modulo). There are more possible combinations of tokens but some rules
still need to be rewritten to resolve correctly. The `generators` are not
included properly in the language, and more rules should be written to
accomodate them. Every lowercase letter is now free to use to include more
operations, operators, etc... There is only ONE parser left, instead
of them being split in three different grammars like before:

- `P()` is the generic object to write/study patterns.

- `P` is named `proto` (for prototype) in the source code.

I also made some cosmetic changes to the Sardine boot process in order to
show more about the user configuration at boot time. It gives visual
feedback about important information to check everytime you start a
session (which MIDI port is toggled? is it deferred or not?), etc...