Pre-merge experimental branch `sysex` into main.
It is time to close this branch to start fresh. This branch was aimed at
preparation of a gig in Lorient (13/10). Focus has now shifted from
performance to enhancing the installation process, the documentation
website and to bring structural additions to Sardine. Closing now seems
like a good option to make the work process clearer for contributors.