Sneaky update of div and speed behavior.
These features are part of what caused such a big trouble three days
ago. They were indexed on a value that was way too low and was putting
the system under too much stress. By indexing div on iterators, we get
the same mechanism at variable speed. This is much better and will not
cause so much trouble :)