Fix fishery hanging on KeyboardInterrupt with LinkClock
I've ported the asyncio.Runner class from 3.11 to make sure that all tasks
are cancelled when the console is exiting. Handlers that use threads should
still remain non-daemonic so they have a proper avenue for handling
cleanup. The FishBowl also now triggers a stop event on deletion in case
there are leftover handlers.