Add multiple aliases for MIDI players keyword parameters: 'chan', 'val', 'dur', etc...
These are really helpful to save a few characters when improvising :)
They were also a personal favorite of the main Sardinist out there, who
was a little bit lost when he tried V2 for the first time this
afternoon.

Also removing a debug/test print statement that was floating out there.