Disable precision event loops
Reference: https://peps.python.org/pep-0564/#annex-clocks-resolution-in-python
For unix-based systems, this change doesn't really matter because the
default monotonic() time already has a precision of ~81ns, equivalent to
perf_counter(). For Windows systems, precision event loops significantly
improves time resolution, but we are still limited to the sleep precision
of Windows, that being ~15ms. In this case, smaller resolutions will
interfere with the ahead-of-time execution done by the event loop,
which might worsen accuracy.