Fixing euclidian rhythm function
Previously, the euclidian rhythm function for use in the mininotation
was completely bugged. It required that the input pattern length was at
least as large as the number of steps in the pattern. This is no longer
the case. The initial pattern will be extended to fit the number of
steps we want to map the euclidian rhythm over.

I've also included a new alias for the euclid() idiom: eu().