Adding the "send_raw" function to the OSC handler.
This function is needed to send a raw (non-sorted, non key-value) OSC
message in some situations (custom audio engines, other tools). Not
having it on the public side was breaking convention. This function is
just an alias of the _send method but this is already a step in the
right direction.