Adding _defaults dict for OSC, MIDI and Dirt Handlers
Adding a self._defaults to every base handler so that the user can store
default values for them (e.g. apply 'speed'=2 on every D or ZD). This
shouldn't cause too much trouble for the moment but some work is needed
to sanitize the input and for some reason, the defaults do not apply for
MIDI patterns...