[BREAKING CHANGE] Refactoring of the Sardine Software Architecture
This is a shallow change but it will have deep implications on the structure of the Sardine library:
- fishery has been renamed to sardine.
- sardine has been renamed to sardine_core.

I'm pretty sure that everything has been renamed or updated to reflect
these changes but I might have missed something in the transition. It
will gradually be corrected in the following commits. The documentation
has been updated to reflect these changes as well. The sardine command
is now very powerful:

- sardine: will start the REPL (previously: fishery)
- sardine web: will start the REPL + web editor (previously fishery web)
- sardine config: will start the configuration tool (previously
  sardine-config)
- sardine config-superdirt: will open the superdirt configuration tool
  (previously sardine-config-superdirt)
- sardine config-python: will open the python configuration tool
  (previously sardine-config-python)

The VSCode package should be updated as well to link to the new command!