Complete Replacement of the Sardine Web Editor
The old editor, written in Svelte, was getting increasingly buggy. This
commit is replacing Sardine Web by Sardine Web II, written using
Typescript, Tailwind and a default Vite project bundler. The two
applications have the same features. The difference is mostly cosmetic.
This rewrite was the occasion to rewrite some of the logic behind
highlighting, evaluation, file management, etc...

- Easier to install (less dependencies, less garbage)
- Better performance (lighter, more efficient)
- Files are stored in the browser localStorage
- Log window disabled until a fix is found.
- Multi-line evaluation multi-statement eval fixed
- Settings menu to switch to Vim Mode and change font
- Help Files are served like any other file!

Another commit will come later tonight to document the changes and to
add some quality of life features to the editor.