Fix: arbitrary callable functions for midi send
The midi sender allows for arbitrary callable functions as parameters.
However, the resolved callables were added to the pattern that ends up
being passed to the method in charge of triggering midi notes, causing
an error.

This fix makes it possible to use any arbitrary callable while dropping
the result: it cannot be used meaningfully to alter the midi message.